+++ src/exp/inc/src/pwr_version.h
@@ -1,6 +1,6 @@
 /*
  * ProviewR   Open Source Process Control.
- * Copyright (C) 2005-2024 SSAB EMEA AB.
+ * Copyright (C) 2005-2026 SSAB EMEA AB.
  *
  * This file is part of ProviewR.
  *
@@ -74,13 +74,13 @@
 #define SET_VERSION(a, b, c, d) ((pwr_tVersion)((a << 24) + (b << 16) + (c << 8) + d))
 #endif
 
-#define pwrv_cBuildTimeStr "01-MAR-2024 12:00:00"
+#define pwrv_cBuildTimeStr "09-MAR-2026 13:00:05"
 
-#define pwrv_cPwrVersionStr "V6.1.3"
-#define pwrv_cPwrVersion SET_VERSION('V', 6, 1, 3)
+#define pwrv_cPwrVersionStr "V7.0.0"
+#define pwrv_cPwrVersion SET_VERSION('V', 7, 0, 0)
 
-#define pwrv_cWbdbVersionShortStr "V61"
-#define pwrv_cPwrCopyright "Copyright (C) 2005-2024 SSAB EMEA AB"
+#define pwrv_cWbdbVersionShortStr "V7"
+#define pwrv_cPwrCopyright "Copyright (C) 2005-2026 SSAB EMEA AB"
 
 #if (pwr_dHost_byteOrder == pwr_dLittleEndian)
 #define PWRV_VERSION_MAJOR (0xFF & (pwrv_cPwrVersion >> 8))