+++ src/lib/co/src/co_dbs.h
@@ -1,7 +1,7 @@
 #include "co_pdr.h"
 /**
  * ProviewR   Open Source Process Control.
- * Copyright (C) 2005-2024 SSAB EMEA AB.
+ * Copyright (C) 2005-2026 SSAB EMEA AB.
  *
  * This file is part of ProviewR.
  *